peso de int para double

+++ include/produto_alimenticio.hpp
@@ -16,7 +16,7 @@ class ProdutoAlimenticio : public Produto {
     
     // Construtor
     ProdutoAlimenticio(const std::string &nome, double preco, int quantidade, bool vegano,
-                       double peso, const std::string &dataValidade);
+                       int peso, const std::string &dataValidade);
 
     /*
      * Imprime as informações do produto alimentício através da função imprimir_informacoes.
